Refactor : moved TokenSearchUtils to token-bridge-sdk
